refactor: move replace authority interface builder
